fixed some issues with gcc installation, now we support Ubuntu 12.04, 14.04 and 16.04: 1. fixed the problem with GMP mismatch with MPFR. 2. automized the make -j option based on

+++ extern/install-gcc-4.5.3.sh
@@ -5,6 +5,16 @@
 
 ### Install GMP
 sudo apt-get install --yes libgmp-dev
+
+if ( ls /usr/include/gmp.h &> /dev/null ); then
+  echo "SUCCESS: found /usr/include/gmp.h"
+elif ( ls /usr/include/x86_64-linux-gnu/gmp.h &> /dev/null ); then
+  echo "SUCCESS: found /usr/include/x86_64-linux-gnu/gmp.h, setting the symlink for /usr/include/gmp.h"
+  sudo ln -s /usr/include/x86_64-linux-gnu/gmp.h /usr/include/gmp.h
+else
+  echo "ERROR: could not fine gmp.h, consider using install-gmp-6.1.1.sh script."
+  exit 2
+fi
 # ./install-gmp-6.1.1.sh
 
 ### Install MPFR
@@ -43,17 +53,27 @@ sudo apt-get install --yes libc6-dev-i386
 sudo ln -s /usr/lib/x86_64-linux-gnu /usr/lib64
 
 ### build
-make -j 12
+make -j $(nproc)
 sudo make install
 cd -
 
-### switch gcc/g++ version
-sudo rm /usr/bin/gcc
-sudo ln -s /usr/bin/gcc-4.5 /usr/bin/gcc
-sudo rm /usr/bin/g++
-sudo ln -s /usr/bin/g++-4.5 /usr/bin/g++
-
-
 rm -rf ${SOURCE_DIR}
 rm -rf ${OBJ_DIR}
+
+### switch gcc/g++ version
+if [ ls /usr/bin/gcc-4.5 &> /dev/null ]; then
+  echo "SUCCESS: gcc-4.5 was installed, setting up the symlink"
+  sudo rm /usr/bin/gcc
+  sudo ln -s /usr/bin/gcc-4.5 /usr/bin/gcc
+else
+  echo "ERROR: gcc-4.5 was NOT installed!"
+fi
+
+if [ ls /usr/bin/g++-4.5 &> /dev/null ]; then
+  echo "SUCCESS: g++-4.5 was installed, setting up the symlink"
+  sudo rm /usr/bin/g++
+  sudo ln -s /usr/bin/g++-4.5 /usr/bin/g++
+else
+  echo "ERROR: g++-4.5 was NOT installed!"
+fi
